refactor: CommaSplitter 구현
