+++ src/test/java/ladder/util/CommaSplitterTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,8 +9,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 public class CommaSplitterTest {
 
     @Test
-    @DisplayName("null or empty 문자열")
-    void testSplitBlank() {
+    @DisplayName("null or empty split 시 예외")
+    void testSplitBlankString() {
         assertThatThrownBy(() -> CommaSplitter.split("")).isInstanceOf(IllegalArgumentException.class);
         assertThatThrownBy(() -> CommaSplitter.split(" ")).isInstanceOf(IllegalArgumentException.class);
         assertThatThrownBy(() -> CommaSplitter.split(null)).isInstanceOf(IllegalArgumentException.class);
@@ -18,8 +18,8 @@ public class CommaSplitterTest {
 
     @Test
     @DisplayName("정상 실행")
-    void testSplit() {
-        final String str = "a,b,c";
-        assertThat(CommaSplitter.split(str)).containsExactly("a", "b", "c");
+    void testSplitString() {
+        String[] array = {"a","b","c"};
+        assertThat(CommaSplitter.split(String.join(",", array))).containsExactly(array);
     }
 }
